This pull request merges in the current set of internal changes into
the SFrame repo.

- [Sphinx-to-zero] Getting sphinx warnings down to zero (minus mxnet).

- Refactor of item_similarity.

- Added DASSERT_DELTA to NDEBUG section.

- Fix graphlab/sframe python log configuration
  Remove use of Queuehandler in root logger

- Fix to make random_split_by_user determistic.
  Fixes https://github.com/dato-code/documents/issues/3612


- Fix a race in libfault publish socket
  Add timeout to distributed remote call (timeout after 60 failed ping)
  Add exception handling to distributed_call and distributed_exec
  Re-enable dml make sgraph test

+++ oss_src/graphlab/util/bitops.hpp
@@ -123,7 +123,7 @@ static inline void flip_bit(T& x, unsigned int bit, _ENABLE_IF_UINT(T))
  * \param n_bits Index of the bit to flip.
  */
 template <typename T>
-static inline T bit_mask(unsigned int n_bits, _ENABLE_IF_UINT(T)) {
+static inline T bit_mask(size_t n_bits, _ENABLE_IF_UINT(T)) {
   return (T(1) << n_bits) - 1;
 }
 
@@ -135,7 +135,7 @@ static inline T bit_mask(unsigned int n_bits, _ENABLE_IF_UINT(T)) {
  * \param index_end   End of interval.
  */
 template <typename T>
-static inline T bit_mask(unsigned int index_begin, unsigned int index_end, _ENABLE_IF_UINT(T)) {
+static inline T bit_mask(size_t index_begin, unsigned int index_end, _ENABLE_IF_UINT(T)) {
   return bit_mask<T>(index_begin) ^ bit_mask<T>(index_end);
 }
